requirements.txt
pyrogram 
tgcrypto
wget
redis
youtube-search-python
pillow
requests
yt_dlp
speedtest-cli
